+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 
 - Upgrade dependencies
 - Add Embroider support
-- Prefer `keyup` over `keydown` where possible
+- Where possible, `keyup` is [preferred](https://github.com/ember-template-lint/ember-template-lint/blob/master/docs/rule/no-down-event-binding.md) over `keydown`
 
 ## 2.2.8